genai/streamlit/requirements.txt
streamlit 
vertexai
google-cloud-aiplatform
llama-index
langchain
PyPDF2
duckdb==1.1.0
google-cloud-secret-manager
pandas
